fix bugs testing on hardware

+++ src/qibocal/plots/calibrate_qubit_states.py
@@ -70,7 +70,7 @@ def qubit_states(folder, routine, qubit, format):
 
         for i, model in enumerate(models_name):
             y_test = _bytes_to_np(parameters.df.iloc[i]["y_test"], np.int64)
-            y_pred = _bytes_to_np(parameters.df.iloc[i]["y_pred"], np.int64)
+            y_pred = _bytes_to_np(parameters.df.iloc[i]["y_pred"], np.float)
             predictions = _bytes_to_np(parameters.df.iloc[i]["predictions"], np.int64)
             # Evaluate the ROC curve
             fpr, tpr, _ = roc_curve(y_test, y_pred)
